Reaction mechanism of titania
This mechanism was first introduced by Livage and Sanchez in the year 1992. R stands for CH(CH3)2. Step 1 shows the reaction. . In step 2 titanium cations suffers nucleophilic addition of the alcoholic OH- group. In step 3, a transfer occurs, without transition state, between the positively charged C2H5 group and the negatively charged OR group. In 
step 4, the separation of the positively charged OR ligand occurs. As a result of these reactions, the OR group is replaced by an OH group, forming titanium hydroxides which undergo condensation. All the steps above occur simultaneously in a dynamic process, promoting the formation of nanoparticles through polycondensation of three
dimensional tetrahedral centers formed by titanium atoms.